解决了大数存储问题
这里的超大数用string类型存储，然后用string 数组一一取出，记得string中存储的字符型号比原来的字符大48.

+++ 1002.h
@@ -2,7 +2,7 @@
 #include <string>
 #include <stack>
 using namespace std;
-void digitalToPinyin(long long n);
+void digitalToPinyin(int n);
 /*
 class Int
 {
